fix(core/ui): resolve all compilation errors in design-system components

+++ core/ui/src/main/java/app/otakureader/core/ui/components/NeonSlider.kt
@@ -18,7 +18,9 @@ import androidx.compose.ui.geometry.CornerRadius
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.*
+import androidx.compose.ui.graphics.drawscope.Stroke
 import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.layout.onSizeChanged
 import androidx.compose.ui.unit.dp
 import kotlin.random.Random
 
@@ -44,7 +46,7 @@ fun NeonSlider(
             Sparkle(
                 offset = sparkleRandom.nextFloat(),
                 size = 2f + sparkleRandom.nextFloat() * 3f,
-                phase = sparkleRandom.nextFloat() * 1000
+                phase = (sparkleRandom.nextFloat() * 1000).toLong()
             )
         }
     }
